+++ rasa/nlu/training_data/formats/markdown.py
@@ -1,6 +1,7 @@
 import logging
 import re
 import typing
+from collections import OrderedDict
 from typing import Any, Text, Optional, Tuple, List, Dict
 
 from rasa.nlu.training_data.formats.readerwriter import (
@@ -211,25 +212,31 @@ class MarkdownWriter(TrainingDataWriter):
     def _generate_training_examples_md(self, training_data: "TrainingData") -> Text:
         """Generates markdown training examples."""
 
-        # Sort training data by intent
-        training_examples = sorted(
-            [e.as_dict_nlu() for e in training_data.training_examples],
-            key=lambda k: k[MESSAGE_INTENT_ATTRIBUTE],
-        )
-        md = ""
-        for i, example in enumerate(training_examples):
-            intent = training_examples[i - 1][MESSAGE_INTENT_ATTRIBUTE]
-            if i == 0 or intent != example[MESSAGE_INTENT_ATTRIBUTE]:
-                md += self._generate_section_header_md(
-                    INTENT,
-                    example[MESSAGE_INTENT_ATTRIBUTE],
-                    example.get(MESSAGE_RESPONSE_KEY_ATTRIBUTE, None),
-                    i != 0,
-                )
+        training_examples = OrderedDict()
 
-            md += self._generate_item_md(self._generate_message_md(example))
+        # Sort by intent while keeping basic intent order
+        for example in [e.as_dict_nlu() for e in training_data.training_examples]:
+            intent = example[MESSAGE_INTENT_ATTRIBUTE]
+            training_examples.setdefault(intent, [])
+            training_examples[intent].append(example)
 
-        return md
+        # Don't prepend newline for first line
+        prepend_newline = False
+        lines = []
+
+        for intent, examples in training_examples.items():
+            section_header = self._generate_section_header_md(
+                INTENT, intent, prepend_newline=prepend_newline
+            )
+            lines.append(section_header)
+            prepend_newline = True
+
+            lines += [
+                self._generate_item_md(self._generate_message_md(example))
+                for example in examples
+            ]
+
+        return "".join(lines)
 
     def _generate_synonyms_md(self, training_data: "TrainingData") -> Text:
         """Generates markdown for entity synomyms."""
@@ -278,32 +285,26 @@ class MarkdownWriter(TrainingDataWriter):
 
     @staticmethod
     def _generate_section_header_md(
-        section_type: Text,
-        title: Text,
-        subtitle: Optional[Text] = None,
-        prepend_newline: bool = True,
+        section_type: Text, title: Text, prepend_newline: bool = True
     ) -> Text:
         """Generates markdown section header."""
 
         prefix = "\n" if prepend_newline else ""
-        subtitle_suffix = (
-            f"{RESPONSE_IDENTIFIER_DELIMITER}{subtitle}" if subtitle else ""
-        )
-        return prefix + "## {}:{}{}\n".format(
-            section_type, encode_string(title), encode_string(subtitle_suffix)
-        )
+        title = encode_string(title)
+
+        return f"{prefix}## {section_type}:{title}\n"
 
     @staticmethod
     def _generate_item_md(text: Text) -> Text:
         """Generates markdown for a list item."""
 
-        return "- {}\n".format(encode_string(text))
+        return f"- {encode_string(text)}\n"
 
     @staticmethod
     def _generate_fname_md(text: Text) -> Text:
         """Generates markdown for a lookup table file path."""
 
-        return "  {}\n".format(encode_string(text))
+        return f"  {encode_string(text)}\n"
 
     def _generate_message_md(self, message: Dict[Text, Any]) -> Text:
         """Generates markdown for a message object."""